client-api: Rename server module to admin
It should be easier to find since it contains the administration
endpoints and matches the namespace in the spec.
