+++ crates/ruma-client-api/src/admin.rs
