+++ crates/ruma-client-api/src/admin/get_user_info.rs
